Commented the Engine

+++ src/main/java/helix/game/GameObject.java
@@ -6,33 +6,96 @@ import helix.game.objects.alarm.Event;
 import helix.utils.math.Point;
 import helix.utils.math.Vector2;
 
+/**
+ * GameObject that stores basic stuff for Game Objects,
+ * such as {@link Alarm}s, a position, direction, and access to {@link Data}
+ * @author bmeachem
+ *
+ */
 public abstract class GameObject {
-	public static int ID_NEXT = 0;
+	/**
+	 * Next ID to be assigned
+	 */
+	public static long ID_NEXT = 0;
 
+	/**
+	 * ID of the object
+	 */
 	public final int id;
 	
+	/**
+	 * Alarms to be used for timed events
+	 */
 	private Alarm[] alarm;
+	
+	/**
+	 * Whether or not this object should be disposed
+	 */
 	private boolean shouldDispose;
 
+	/**
+	 * Position of the object
+	 */
 	private Point pos;
+	
+	/**
+	 * Direction the object travels in
+	 */
 	private Vector2 direction;
 
+	/**
+	 * Data of the application
+	 */
 	private final Data data;
 	
+	/**
+	 * To be overridden as necessary.
+	 * Ran before {@link GameObject#step}
+	 * @param delta - Time since last frame (seconds)
+	 * 
+	 * @see {@link GameObject#step}, {@link GameObject#postStep}
+	 */
 	protected void preStep(float delta) {}
+	
+	/**
+	 * Basic step event for {@link GameObject}
+	 * @param delta - Time since last frame (seconds)
+	 * 
+	 * @see {@link GameObject#preStep}, {@link GameObject#postStep}
+	 */
 	public abstract void step(float delta);
+	
+	/**
+	 * To be overridden as necessary.
+	 * Called after {@link GameObject#step}
+	 * <br>
+	 * Mainly used for cleanup
+	 * @param delta - Time since last frame (seconds)
+	 * 
+	 * @see {@link GameObject#preStep}, {@link GameObject#step}
+	 */
 	protected void postStep(float delta) {}
 	
+	/**
+	 * Create a new GameObject
+	 * @param data - All the Game Data
+	 * @param pos - {@link Point} to spawn the Object at
+	 */
 	public GameObject(Data data, Point pos) {
 		this.pos = pos;
 		this.data = data;
 		this.direction = new Vector2(0, 0);
 		this.initAlarms();
 		
-		this.id = ID_NEXT++;
+		this.id = (int)(ID_NEXT++);
 		data.objects.add(this);
 	}
 	
+	/**
+	 * Initialize each alarm
+	 * 
+	 * @see {@link Alarm}, {@link Alarm#ALARM_COUNT}
+	 */
 	private void initAlarms() {
 		alarm = new Alarm[Alarm.ALARM_COUNT];
 		for(int i = 0; i < Alarm.ALARM_COUNT; i++) {
@@ -40,12 +103,20 @@ public abstract class GameObject {
 		}
 	}
 
+	/**
+	 * Update the game object
+	 * @param delta - Time since last frame (seconds)
+	 */
 	public final void update(float delta) {
 		preStep(delta);
 		step(delta);
 		postStep(delta);
 	}
 	
+	/**
+	 * Update the alarms
+	 * @param delta - Time since last frame (seconds)
+	 */
 	public final void updateAlarms(float delta) {
 		for(Alarm alarm : alarm) {
 			alarm.update(delta);
@@ -53,6 +124,11 @@ public abstract class GameObject {
 	}
 	
 	@SuppressWarnings("unchecked")
+	/**
+	 * Find the first instance of some {@link GameObject}
+	 * @param searchClass - Type of class to search for
+	 * @return - the first instance of searchClass or null if none is found
+	 */
 	public final <T extends GameObject> T find(Class<T> searchClass) {
 		for(GameObject object : data.objects) {
 			if(searchClass.isInstance(object))
@@ -62,6 +138,11 @@ public abstract class GameObject {
 	}
 	
 	@SuppressWarnings("unchecked")
+	/**
+	 * Find the first instance of some {@link GameObject}
+	 * @param searchClass - Type of class to search for
+	 * @return - the nearest instance of searchClass or null if none is found
+	 */
 	public final <T extends GameObject> T findNearest(Class<T> searchClass) {
 		GameObject current = null;
 		for(GameObject object : data.objects) {
@@ -81,28 +162,55 @@ public abstract class GameObject {
 			return (T)current;
 		else return null;
 	}
-	
+
+	/**
+	 * Get the distance to another {@link GameObject}
+	 * @param other - Other Object to get distance to
+	 * @return distance (float)
+	 */
 	public final float distTo(GameObject other) {
 		return this.getPos().sub(other.getPos()).length();
 	}
 
+	/**
+	 * Move towards {@link GameObject#direction} at a given speed
+	 * @param speed (px/frame)
+	 */
 	public final void move(float speed) {
 		this.setPos(getPos().add(direction.multiply(speed)));
 	}
 	
+	/**
+	 * Move towards a {@link Point} at a given speed
+	 * @param point - Point to move towards
+	 * @param speed - Speed to move at (px/frame)
+	 */
 	public final void moveTo(Point point, float speed) {
 		this.direction = point.sub(this.getPos()).toVector2().getUnitVector();
 		this.move(speed);
 	}
-	
+
+	/**
+	 * Move towards a {@link GameObject} at a given speed
+	 * @param target - Target to move towards
+	 * @param speed - Speed to move at (px/frame)
+	 */
 	public final void moveTo(GameObject target, float speed) {
 		this.moveTo(target.getPos(), speed);
 	}
-	
+
+	/**
+	 * Move towards a {@link Point} at {@link Constants#DEFAULT_SPEED}
+	 * @param point - Point to move towards
+	 */
 	public final void moveTo(Point point) {
 		this.moveTo(point, Constants.DEFAULT_SPEED);
 	}
-	
+
+	/**
+	 * Move towards a {@link GameObject} at {@link Constants#DEFAULT_SPEED}
+	 * @param target - Target to move towards
+	 */
 	public final void moveTo(GameObject target) {
 		this.moveTo(target, Constants.DEFAULT_SPEED);
 	}
